Refactor import paths in CombatSystem and types for improved organization and clarity
